requirements.txt
line-bot-sdk
gunicorn
flask
tensorflow
matplotlib
numpy
keras
ipython
Pillow